- Remove API Assessment Service - Fix content helper bug
