+++ DFC.App.DiscoverSkillsCareers.Models/Assessment/DysacAssessment.cs
@@ -6,6 +6,7 @@ namespace DFC.App.DiscoverSkillsCareers.Models.Assessment
     public class DysacAssessment : DocumentModel
     {
         public override string? PartitionKey { get; set; } = "/Assessment";
+
         public IEnumerable<ShortQuestion> Questions { get; set; }
 
         public string? AssessmentCode { get; set; }
